% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_model.R
\name{bay_voronoi}
\alias{bay_voronoi}
\title{Computes the bayesian posterior distribution using a voronoi tesselation}
\usage{
bay_voronoi(df_inter, prior = NULL, var_prior = NULL,
  var_final = "proba_final", filter_null = TRUE)
}
\arguments{
\item{df_inter}{a \code{data.frame} which indicates for each voronoi x grid
intersection the relative area of the intersection in the corresponding
voronoi; should be either produced by or have the same format as the
output of \code{\link{inter_grid_voronoi}}.}

\item{prior}{a \code{data.frame}; should be either produced by or have
the same format as those produced by the prior computation functions
of \code{\link{mobicount}}. Defauts to \code{NULL}, in which case a prior
equal to one is used (uniform).}

\item{var_prior}{character; name of the variable which contains the
prior distribution; ignored if \code{prior = NULL}.}

\item{var_final}{character; name of the variable which should contain
the posterior distribution.}

\item{filter_null}{logical; should voronoi x grid intersections with
a null posterior probability be deleted ? Defauts to \code{TRUE}.}
}
\value{
a \code{data.frame} with two columns : an ID for each
voronoi x grid intersection formatted as "voronoi_ID:grid_id",
and the posterior distribution.
}
\description{
This function implements the bayesian model for mobile events detection,
assuming that no information on the MNO's antennas coverage is available.
A voronoi tesselation is thus used to approximate the area covered by
each antenna. If available, prior information over tiles can be used
to improve the quality of the detection.
}
\details{

}
\examples{
\dontrun{
# Import prior
prior_bdtopo <- readr::read_csv(~/bdtopo_france.csv)
# Import voronoi x grid intersections table
inter_df <- readr::read_csv(~/inter_grid_voro.csv)
# Compute bayesian posterior distribution
bay_df_bdtopo <- bay_voronoi(inter_df, prior_bdtopo, "proba_bdtopo")
}
}
