% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_computation.R
\name{prior_combine}
\alias{prior_combine}
\title{Combines multiple priors}
\usage{
prior_combine(priors, weights)
}
\arguments{
\item{priors}{a list of \code{data.frame}s, one for each prior.
See 'Details' for required structure}

\item{weights}{numeric vector of same length as \code{priors}; weights used
to compute the linear combination of priors.\cr
Must sum to one for the output to be a probability distribution.}
}
\value{
a \code{data.frame} with two columns : tiles ID and a prior
distribution over tiles.
}
\description{
Combines multiple prior distributions over a same grid
by computing a new distribution as a linear combination
}
\details{
\code{data.frame}s in \code{priors} should have the same structure
as those produced by \link{prior_building} and \link{prior_res_pop} :\cr
a first column indicating tiles IDs, and a second column giving the
distribution of the prior on the grid.
}
\examples{
\dontrun{
# Import priors
prior_bdtopo <- prior_building(paths_shp = "~/bdtopo_shp", grid = grid_500_france)
prior_rfl <- prior_res_pop(rfl = rfl11)
# Include inputs in a list
list_priors <- list(prior_bdtopo, prior_rfl)
# Define weights to be used (same order as the list of priors)
weights = c(0.3, 0.7)
# Combine priors
prior_comb <- prior_combine(priors = list_priors, weights = weights)
}
}
