% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_computation.R
\name{prior_res_pop}
\alias{prior_res_pop}
\title{Computes prior from resident population}
\usage{
prior_res_pop(rfl, indiv_id_var = "DIRNOSEQ", n_indiv_var = "nbpersm",
  x_y_var = c("x", "y"))
}
\arguments{
\item{rfl}{a \code{data.frame} of class \code{\link[data.table]{data.table}}
containing fiscal localized incomes data. If a simple \code{data.frame}
is provided, it is automatically converted to a \code{data.table}.}

\item{indiv_id_var}{character; name of the variable indicating household
unique identifier in \code{rfl}.}

\item{n_indiv_var}{character; name of the variable indicating household
sizes (number of individuals) in \code{rfl}.}

\item{x_y_var}{character vector of length 2; names of the variables
indicating x and y coordinates of the household in \code{rfl}.}
}
\value{
a \code{data.frame} of class \code{\link[data.table]{data.table}}
with two columns : tiles ID and a prior distribution over tiles.
}
\description{
Computes a prior distribution over a grid using resident population data
from fiscal localized income sources (e.g. RFL or FiLoSoFi data in France).
}
\details{
If the output \code{data.frame} of this function is to be used as
in \code{\link{}}, x and y coordinates in input data (\code{rfl}) should
be converted beforehand in Lambert-93 (ESPG:2154).
\cr \cr
In the output \code{data.frame}, tiles ID are formatted
as "x_centroid/100:y_centroid/100" in order to match the formatting of
grids created by \code{\link{create_grid}}.
}
\examples{
\dontrun{
library(sas7bdat)
# Import data from fiscal localized data in 2014
rfl_df <- read.sas7bdat("path_to_filosofi_2014/menages14.sas7bdat")
# Compute prior using resident population
res_prop_prior <- prior_res_pop(rfl = rfl_df, indiv_id_var = "IDENTIFIANT")
}
}
